Global stats effecs and destroyable Bonfire
- Global stat upgrades now take effect on weapons: Amount, cooldown, speed, and others
- Bonfire behavior so the player has to destroy them to get the collectable. GD choice as to no one moves towards a coin hah
- Tweaked some values, gamefeel is beter now

+++ Assets/Scripts/WeaponBase.cs
@@ -10,13 +10,13 @@ public class WeaponBase : MonoBehaviour
 
 	#region PRIVATE_VARIABLES
 
-	public int weaponLevel;
-	public int maxLevel;
-	public int baseDamage;
-	public int amount;
-	public int pierce;
-	public int poolLimit;
-	public int rarity;
+	public float weaponLevel;
+	public float maxLevel;
+	public float baseDamage;
+	public float amount;
+	public float pierce;
+	public float poolLimit;
+	public float rarity;
 	public float speed;
 	public float area;
 	public float duration;
